// api/generated/model/gymChoiceInfoDto.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * LgymApi.Api
 * OpenAPI spec version: 1.0
 */
import type { LastTrainingGymInfoDto } from './lastTrainingGymInfoDto';

export interface GymChoiceInfoDto {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  _id?: string | null;
  lastTrainingInfo?: LastTrainingGymInfoDto;
}